tools/grafana_info/requirements.txt
grafana-api==0.2.4
simplejson==3.16.0
pyyaml==3.13
